refactor(diagnosis): six files in reading order, each assuming only those above
The comments referenced each other in both directions. Measured over the
package, the file graph held nine mutual dependencies, so no ordering of the
files let a reader meet every term after its definition. Not a bad order: none
existed. That is why the previous two passes, which fixed accuracy and then
length, both left the package hard to read. Every sentence can be true and
short and the package still cannot be read in any order.

Eleven files become six, and they build in one direction:

    measure.go   a reading, and the calculations that summarise a series
    window.go    where readings accumulate and age out
    declare.go   what a caller writes: capabilities, instruments, the table
    judge.go     thresholds, the latch, and how fired signals are ordered
    engine.go    builds it from the table and runs one tick
    suite.go     generates a conformance suite for a caller's table

A file may use a term declared in itself or above it, never below. Where a
comment reached forward, the fact was usually in the wrong file rather than
wrong, so six of them moved to the later file that owns them: three refusals
and the capacity-zero bypass into validate, the Reset caller into Observe.

Declarations were reordered inside files for the same reason. In judge.go a
reader met a three-row table naming Update, ReleaseAfter and Reset fifty lines
before they were declared; Marks was declared above the Polarity type it has a
field of.

One vocabulary per concept: "fold" and "folded" are gone in favour of reduce,
and "outcome" in favour of State. Abstract terms now name a concrete instance,
so a capability is "PSI is not available on every kernel" rather than "a
startup fact about the environment".

Two errors found while rewriting. NoInstrument claimed the latch "releases on
the demote clock rather than holding", which inverts ReleaseAfter: it holds
first and releases only once the span has passed, and it contradicted its
sibling constant. Severity's account of a zero capacity covered the rising
polarity and not the falling one, where the same collapse happens over a
negative fire mark.

The environment question this raised is now answered where it is asked:
NewEngine takes no Environment and builds a window for every instrument,
capable or not. Capability is re-read on every tick inside Observe, so one
table runs on boxes that differ and a change takes effect on the next tick.

Comment-to-code goes 0.84 to 0.51.
